snapshot of board merge
a point about the flipped LEDs with a hole:
it's mounted in reverse with a hole as a makeshift lightguard,
...because maybe.
There are actual LEDs with an lens attached (see knightbright),
but regular ones work as well obviously.
Also it allows for text to be silk'd right next to it, which is sweet.